+++ app/routes/_index.tsx
@@ -1,7 +1,7 @@
 import { useLoaderData } from "@remix-run/react";
 import { UserCategory, UserData } from "~/utils/types";
 import { useEffect, useState } from "react";
-import { loader } from "~/loaders/solicitar-pagamento-loader";
+import { loader } from "~/handlers/solicitar-pagamento-loader";
 import { getTelegramUserInfo } from "~/utils/users";
 import telegramInit from "~/hooks/telegramInit";
 import { ButtonsListWithPermissions } from "~/components/CommonButtons";
