+++ app/routes/grupos-de-trabalho.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 import { GrupoTrabalho, UserCategory, UserData } from "~/utils/types";
 import Unauthorized from "~/components/Unauthorized";
 import { isAuth } from "~/hooks/isAuthorized";
-import { loader } from "../loaders/solicitar-pagamento-loader";
+import { loader } from "../handlers/solicitar-pagamento-loader";
 import { getTelegramUserInfo } from "~/utils/users";
 export { loader }
