+++ app/routes/solicitar-pagamento.tsx
@@ -14,8 +14,8 @@ import ValorInput from "~/components/Forms/ValorInput";
 import { UserCategory, UserData } from "../utils/types";
 import { Project } from "~/utils/types";
 
-import { loader } from "~/loaders/solicitar-pagamento-loader";
-import { action } from "~/loaders/solicitar-pagamento-action";
+import { loader } from "~/handlers/solicitar-pagamento-loader";
+import { action } from "~/handlers/solicitar-pagamento-action";
 import { getTelegramUserInfo } from "~/utils/users";
 import { isAuth } from "~/hooks/isAuthorized";
 import Unauthorized from "~/components/Unauthorized";
